pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-juicer-agents"
version = "0.1.0"
description = "A multi-agent data processing system built on AgentScope and Data-Juicer"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "DataJuicer Team" }]
requires-python = ">=3.10, <3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "agentscope>=1.0.10",
    "agentscope-runtime>=1.0.0",
    "jinja2>=3.1.6",
    "langchain-community",
    "psutil>=7.1.3",
    "py-data-juicer>=1.4.0",
    "streamlit>=1.51.0",
    "huggingface-hub<1.0",
    "faiss-cpu>=1.13.1",
    "hf-transfer>=0.1.9",
    "transformers>=4.57.3",
    "oyaml",
]

[project.scripts]
djx = "data_juicer_agents.cli:main"

[project.optional-dependencies]
dev = ["pytest>=6.0", "black", "flake8", "mypy"]
qa = ["qdrant-client>=1.16.2"]

[project.urls]
Homepage = "https://github.com/datajuicer/data-juicer-agents"
Repository = "https://github.com/datajuicer/data-juicer-agents"
Documentation = "https://github.com/datajuicer/data-juicer-agents#readme"

[tool.setuptools]
packages = [
    "data_juicer_agents",
    "data_juicer_agents.agents",
    "data_juicer_agents.core",
    "data_juicer_agents.commands",
    "data_juicer_agents.runtime",
    "data_juicer_agents.tools",
    "data_juicer_agents.tools.op_manager",
    "data_juicer_agents.utils",
    "qa-copilot",
    "qa-copilot.op_manager",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
